Make index.js re-export only
